Catch exceptions in GameEntryBuilder internally
GameEntryBuilder is used as a testing utility class to create GameEntry objects.

There should not be exceptions when using this class because it is invoked by the programmer to create test case objects.

Let's handle exceptions internally to remove the need to catch when using it to build test classes and cases.