Update clear command to ask for confirmation from the user
CommandWindow is created to show a pop up window for user confirmation.
ClearCommandParser is used to determine the type of ClearCommand to be created.

A clear command from the user should not contain additional arguments, while
a clear command initiated by the app after user's confirmation should contain
an extra identifiable argument.